Enables support for user-specified *http.Clients.
Adds an *http.Client to the main Client struct (and automatically
defaults to http.DefaultClient if nil is provided). This permits
the user to set important client attributes such as request
timeouts (since the default is never to time out).

Requires API change from NewClient() to NewClient(nil) or
NewClient(*http.Client).

Tests are updated and pass.